+++ changelog.md
@@ -1,6 +1,12 @@
 # Changelog
 
 ## [Unreleased]
+### features
+- add `getIcon` fn to the scopes of toolbar placeholders.
+
+### refactored
+- improve interfaces and documentation.
+
 ## [1.9.9]
 ### refactored
 - internal interfaces to improve documentation and development.